move helper func from util.go to querier.go
